+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] -
 ### Added
 - DHCP service API calls and service compiler
+- Database migration for the KeepAlived service management.
+- Database migration for the HAProxy service management.
 - Added query data to get Agent host information.
 - HAProxy plugin.
 - Systemctl management API call.